==> 3.4-MOLfiles PyG MoleculeNet.ipynb <==
# MoleculeNet Pytorch Geometric
In this notebook, the file input in the style of MoleculeNet is loaded. </br>
Note that, due to problems with the local environment and pytorch, this notebook was tested and developed to work in google colab. 

pip install rdkit

# Install required packages.
import os
import torch
os.environ['TORCH'] = torch.__version__
print(torch.__version__)

!pip install -q torch-scatter -f https://data.pyg.org/whl/torch-${TORCH}.html
!pip install -q torch-sparse -f https://data.pyg.org/whl/torch-${TORCH}.html
!pip install -q git+https://github.com/pyg-team/pytorch_geometric.git

import pandas as pd
import rdkit
from torch_geometric.io import read_txt_array
from torch_geometric.data import Data

## Load the different .txt docoments

x = read_txt_array('2-Pytorch geometric data/MoleculeNet/MOLfiles/raw/mol_x.txt', sep = ',')
edge_index = read_txt_array('2-Pytorch geometric data/MoleculeNet/MOLfiles/raw/edge_index.txt', sep = ',').t().to(torch.long)
edge_attr = read_txt_array('2-Pytorch geometric data/MoleculeNet/MOLfiles/raw/edge_attr.txt', sep = ',')

## Load the data
Use the pytorch dataloader to load the data

data = Data(x=x, edge_index=edge_index, edge_attr=edge_attr)

Show example of the data

data.edge_stores
